requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
discord.py==2.4.0
discord==2.3.2
fastapi==0.115.6
feedparser==6.0.11
flask==3.1.0
frozenlist==1.4.1
h11==0.14.0
html2text==2024.2.26
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
multidict==6.1.0
pip==24.2
pydantic-core==2.27.1
pydantic==2.10.3
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.1.0
sgmllib3k==1.0.0
sniffio==1.3.1
starlette==0.41.3
typing-extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
Werkzeug==3.1.3
wheel==0.44.0
yarl==1.13.1
